+++ tango/io/vfs/ZipFolder.d
@@ -860,7 +860,7 @@ private:
                     null));
     }
 
-    void mutate_write(uint bytes, const(void)[] src)
+    void mutate_write(size_t bytes, const(void)[] src)
     {
         if( !(bytes == 0 || bytes == IConduit.Eof) )
             this.modified = true;
@@ -1468,7 +1468,7 @@ class ZipSubFolderGroup : VfsFolders
         return result;
     }
 
-    final uint files()
+    final size_t files()
     in { assert( valid ); }
     body
     {
@@ -1480,14 +1480,14 @@ class ZipSubFolderGroup : VfsFolders
         return files;
     }
 
-    final uint folders()
+    final size_t folders()
     in { assert( valid ); }
     body
     {
         return members.length;
     }
 
-    final uint entries()
+    final size_t entries()
     in { assert( valid ); }
     body
     {
@@ -1600,7 +1600,7 @@ class ZipFileGroup : VfsFiles
         return result;
     }
 
-    final uint files()
+    final size_t files()
     in { assert( valid ); }
     body
     {
@@ -1817,7 +1817,7 @@ class EventSeekInputStream : InputStream //, IConduit.Seek
     {
         void delegate()                     close; ///
         void delegate()                     clear; ///
-        void delegate(uint, void[])         read; ///
+        void delegate(size_t, void[])       read; ///
         void delegate(long, long, Anchor)   seek; ///
     }
 
@@ -1898,10 +1898,10 @@ class EventSeekOutputStream : OutputStream //, IConduit.Seek
     ///
     struct Callbacks
     {
-        void delegate()                     close; ///
-        void delegate()                     flush; ///
-        void delegate(uint, const(void)[])  write; ///
-        void delegate(long, long, Anchor)   seek; ///
+        void delegate()                      close; ///
+        void delegate()                      flush; ///
+        void delegate(size_t, const(void)[]) write; ///
+        void delegate(long, long, Anchor)    seek; ///
     }
 
     //alias IConduit.Seek.Anchor Anchor;
